Move pool classification to eth_token and fix PnL data types
Pool classification logic moved from alpha/pool_classification to
eth_token/src/pools/classification.rs as the canonical home. The alpha
crate becomes a thin re-export. Legacy JSON parsing, stdin binary, and
strategy_stats() preset removed; call sites in eth_chain_server now
inline the config explicitly.

PnL scaled values (token_balance, denom_cashflow, native_fee,
native_bribe, marked_token_value_denom, pnl_proxy_denom) changed from
f64 to rust_decimal::Decimal throughout eth_token and eth_token_pnl_store.
SQL columns changed from DOUBLE PRECISION to NUMERIC. Eliminates NaN/Inf
risk and preserves 18-decimal precision for ETH-denominated values.

+++ eth_token/src/pnl/model.rs
@@ -1,3 +1,4 @@
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
@@ -46,12 +47,12 @@ pub struct PnlAddressPositionExport {
     pub movement_count: u64,
     pub token_balance_raw: String,
     pub denom_cashflow_raw: String,
-    pub token_balance: f64,
-    pub denom_cashflow: f64,
-    pub native_fee: f64,
-    pub native_bribe: f64,
-    pub marked_token_value_denom: Option<f64>,
-    pub pnl_proxy_denom: Option<f64>,
+    pub token_balance: Decimal,
+    pub denom_cashflow: Decimal,
+    pub native_fee: Decimal,
+    pub native_bribe: Decimal,
+    pub marked_token_value_denom: Option<Decimal>,
+    pub pnl_proxy_denom: Option<Decimal>,
 }
 
 #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
